rename & move components
